Initial component commit Expand hooks.ts to directory to split out hooks by use
