+++ eventkit_cloud/ui/static/ui/app/components/CreateDataPack/BreadcrumbStepper.tsx
@@ -33,7 +33,7 @@ import EstimateLabel from "./EstimateLabel";
 import PermissionsBanner from "../PermissionsBanner";
 import {userStatusReducer} from "../../reducers/userReducer";
 import {useState} from "react";
-import {useEffectOnMount} from "../../utils/hooks";
+import {useEffectOnMount} from "../../utils/hooks/hooks";
 import {CreatePagePermissionsBanner} from "./CreatePagePermissionsBanner";
 
 export interface JobData {
