+++ eventkit_cloud/ui/static/ui/app/components/MapTools/OpenLayers/MouseWheelZoom.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect} from 'react';
 import {useOlMapContainer, useOlZoom} from "../context/OpenLayersContext";
-import {useEffectOnMount} from "../../../utils/hooks";
+import {useEffectOnMount} from "../../../utils/hooks/hooks";
 import MouseWheelZoom from "ol/interaction/mousewheelzoom";
